Add a fastpath for CreateCycles in memoryStore
Our workload was spending most of its time in CreatesCycle, which was
spending most of its time constructing the PredecessorMap. Because we
don't need to modify the PredecessorMap, we don't really need to make
any copies and can use memoryStore's inEdges to satisfy CreatesCycle,
instead.

This seemed like the least intrusive way to make this change, but let me
know if you have another preferred way to do this.

I also only added this to directed.go because that's all we use, please
let me know if you'd like me to do similar things elsewhere.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>